Auth refactor
1. Refactored AuthContext by removing logic and only keeping state and basic state manipulation inside.
2. Created separate Hook useAuth.js for auth related higher order logic. The hook imports the context for direct state manipulation.
3. Simplified OAuth functions further by creating reusable signInWithProvider and signInWithProviderToken functions.
4. Created some folders for non-organized screens.
5. Created an index.js for simpler context import.
6. Removed a lot of self-explanatory comments and console.logs from auth logic methods
