+++ frontend/src/components/forms/SignInForm.jsx
@@ -2,7 +2,7 @@
 import { useState, useEffect } from "react";
 import { useLocation, useNavigate } from "react-router-dom";
 
-import { useAuth } from "context/AuthContext";
+import { useAuth } from "hooks/useAuth";
 
 function SignInForm() {
   const location = useLocation();
