+++ frontend/src/screens/gdpr/PrivacyPolicyScreen.jsx
